requirements.txt
# ── Web Framework ──────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0

# ── Geospatial ─────────────────────────────────────────────────────
earthengine-api>=0.1.390
geopandas>=0.14.0
shapely>=2.0.0
fiona>=1.9.0
pyproj>=3.6.0
rasterio>=1.3.0

# ── ML ─────────────────────────────────────────────────────────────
torch>=2.0.0
torchvision
torchaudio
segmentation-models-pytorch>=0.3.4
numpy>=1.24.0
scipy>=1.11.0
scikit-image>=0.21.0

# ── Image Processing ───────────────────────────────────────────────
pillow>=10.0.0
matplotlib>=3.7.0

# ── Database ───────────────────────────────────────────────────────
psycopg2-binary>=2.9.0

# ── Utilities ──────────────────────────────────────────────────────
requests>=2.31.0
pandas>=2.0.0
tqdm>=4.65.0
diskcache>=5.6.0
